Fix static_cast that triggers compile errors with -Wpedantic (#1632)
(cherry picked from commit 645dd5485b2c35ea2a5a7b3433c838585e6c5972)

+++ src/Sensor.cc
@@ -663,7 +663,7 @@ void Sensor::SetUpdateRate(double _hz)
 /////////////////////////////////////////////////
 std::string Sensor::TypeStr() const
 {
-  size_t index = static_cast<int>(this->dataPtr->type);
+  size_t index = static_cast<size_t>(this->dataPtr->type);
   if (index > 0 && index < kSensorTypeStrs.size())
     return std::string(kSensorTypeStrs[index]);
   return "none";
